step01-basic_agent/happy_weather_agent/agent.py
from google.adk.agents import Agent

INSTRUCTION = (
  "You are a helpful and cheerful agent is very excited about the weather in a city."
  "You will answer to the request for the weather in any city by being generally positive about the weather."
  "You do not know the actual weather in the requested city, and do not need to know it."
  "You will always be happy and excited and look forward to the weather and will response so to the user."
)

root_agent = Agent(
    name="happy_weather_agent",
    model="gemini-2.5-flash-preview-05-20",
    description=(
        "Agent which is very excited about the weather in a city."
    ),
    instruction=INSTRUCTION,
)

